非同期処理の仕様変更/時間単位をミリ秒で統一/メッセージの型の変更
* 非同期処理の仕様変更
サンプリングと更新を別スレッドで非同期に。それぞれのIntervalを倍数に設定する必要はなくなった
* 時間単位をミリ秒で統一
Intervalは小数型に変更

+++ app/src/main/java/jp/naist/rocatmonitor/androcatmonitorxp/Config.java
@@ -8,8 +8,8 @@ public class Config
 
   public String Host;
   public int Port;
-  public int SampleInterval;
-  public int UpdateInterval;
+  public double SampleInterval;
+  public double UpdateInterval;
 
   public Set<String> IgnorePackages = new HashSet<String>() { { add(ConstValue.THIS_PACKAGE_NAME); } };
